# docs/swagger.yaml
definitions:
...
  controller._ResponseAllTagList:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        description: '`json:"data,omitempty"` //omitempty 该字段为空时忽略'
        items:
          $ref: '#/definitions/models.ResponseTagBrief'
        type: array
      msg: {}
    type: object
...
  controller._ResponseNoDataArea:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      msg: {}
    type: object
  controller._ResponseTagDetailList:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        allOf:
        - $ref: '#/definitions/models.ResponseTagDetail'
        description: '`json:"data,omitempty"` //omitempty 该字段为空时忽略'
      msg: {}
    type: object
  controller.ResponseCodeType:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeParameterInvalid
    - CodeUserNotExist
    - CodeUserExist
    - CodePasswordError
    - CodeServerBusy
    - CodeTokenInvaild
    - CodeTokenEmpty
    - CodeNeedReLogin
    - CodeCommunityIdInvalid
    - CodeVerifyCodeInvaild
    - CodeArticleTitleExisted
    - CodeArticleNotExisted
    - CodeHaveArticleInClass
    - CodeTagExisted
    - CodeTagNotExisted
    - CodeClassNotExisted
    - CodeClassNameExisted
    - CodeCreateCommentFailed
...
  models.DeleteMultiTagParams:
    properties:
      ids:
        description: 待删除的TagID列表
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  models.ResponseTagBrief:
    properties:
      atc_count:
        description: 该标签下包含的文章数量
        type: integer
      name:
        description: 类别名称
        type: string
      tag_id:
        description: 标签ID - 由应用层生成
        example: "0"
        type: string
    type: object
  models.ResponseTagDetail:
    properties:
      atc_count:
        description: 该标签下包含的文章数量
        type: integer
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 标签的描述信息
        type: string
      name:
        description: 类别名称
        type: string
      tag_id:
        description: 标签ID - 由应用层生成
        example: "0"
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  models.TagCreateFormParams:
    properties:
      desc:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  models.UpdateTagParams:
    properties:
      desc:
        type: string
      name:
        type: string
      tag_id:
        example: "0"
        type: string
    required:
    - name
    - tag_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: NgBlog Go博客项目 API 接口文档
  title: NgBlog
  version: "1.0"
paths:
...
  /tag:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除多个Tag
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除多个标签的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.DeleteMultiTagParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 删除多个标签(Tag)的接口
      tags:
      - 标签相关接口
    post:
      consumes:
      - application/json
      description: 通过该接口可以创建标签
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建标签的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.TagCreateFormParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 创建标签(tag)的接口
      tags:
      - 标签相关接口
    put:
      consumes:
      - application/json
      description: 通过该接口可以更新Tag
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新标签的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTagParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 更新标签(Tag)的接口
      tags:
      - 标签相关接口
  /tag/:id:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除单个Tag
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 删除单个标签(Tag)的接口
      tags:
      - 标签相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以获得当前标签的详细信息
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseTagDetailList'
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取标签的详细信息的接口
      tags:
      - 标签相关接口
  /tags:
    get:
      consumes:
      - application/json
      description: 通过该接口可以获得当前的所有标签
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseAllTagList'
      security:
      - ApiKeyAuth: []
      summary: 获取所有标签（简略信息）的接口
      tags:
      - 标签相关接口
swagger: "2.0"